+++ resources/views/external_mentorship_guidelines.blade.php
@@ -1,0 +1,167 @@
+<p>
+    Dear Andelan,
+    We've been unable to match your request on Lenken, please find the guide to requesting external mentorship below:
+
+</p>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
+"http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml" lang="en" xml:lang="en" style="width: 100%; height: 100%;">
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+    <meta name="viewport" content="width=device-width">
+    <!-- This will be the subject of the mail -->
+    <title>External Engagement Rquests</title>
+</head>
+<body style="font-size: 16px; font-family: 'Avenir', 'Avenir Next', Helvetica; padding: 0; margin: 0; color: #666; width: 100%; height: 100%;">
+<!-- Main table wrapper -->
+<table cellpadding="0" cellspacing="0" style="width: 100%; height: 100%; border-collapse: collapse; background-color: #f3f3f3;" valign="top">
+    <tbody>
+    <tr>
+        <td align="center" valign="top">
+            <center>
+                <!-- Container -->
+                <table width="680" style="padding: 20px;">
+                    <tbody>
+                    <tr>
+                        <td style="background-color: #fff; display: block; padding: 25px;">
+                            <!-- Header -->
+                            <table align="center" valign="top" style="border-collapse: collapse; margin-bottom: 15px;">
+                                <tbody>
+                                <tr>
+                                    <th bgcolor="#225be2">
+                                        <table>
+                                            <tbody>
+                                            <tr>
+                                                <!-- Header content -->
+                                                <th width="274" style="padding-right: 16px; padding-left: 16px;">
+                                                    <table>
+                                                        <tbody>
+                                                        <tr>
+                                                            <th width="300">
+                                                                <h3 style="color: #fff; font-weight: 50; text-align: left; font-size: 15px">
+                                                                    External Engagement Requests
+                                                                </h3>
+                                                            </th>
+                                                        </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </th>
+                                                <!-- Header content -->
+
+                                                <!-- Logo -->
+                                                <th width="274" align=right style="padding-right: 16px; padding-left: 16px;">
+                                                    <table>
+                                                        <tbody>
+                                                        <tr>
+                                                            <th>
+                                                                <a href="http://andela.com" target="_blank">
+                                                                    <img src='https://lv.andela.com/images/logo-white.png' height="50px"/>
+                                                                </a>
+                                                            </th>
+                                                        </tr>
+                                                        </tbody>
+                                                    </table>
+                                                </th>
+                                                <!-- Logo -->
+                                            </tr>
+                                            </tbody>
+                                        </table>
+                                    </th>
+                                </tr>
+                                </tbody>
+                            </table>
+                            <!-- Header -->
+
+                            <!-- Body -->
+                            <b>Instructions for Mentees to create a mentorship request</b>
+
+                            <ol type="1">
+                                <li>Create an account on Codementor </li>
+                                <li> Click the Get Help Now button, then click Find a long-term mentor </li>
+                                <li> Enter a brief description of request "[Primary Technologies] for X hrs/day"
+                                    (e.g. React and Ruby on Rails for 2hrs/day) and click Next
+                                </li>
+                                <li>Add relevant categories if needed (5 max).</li>
+                                <li>Looking for someone experienced X, Y, Z technologies.
+                                    I'm a Developer at Andela (*a startup that helps youths in Africa become Software developers*) and
+                                    seeking someone who can commit at least X hours per weekday in the GMT+1 timezone for at least X month(s).
+                                    If you're available on Mon - Fri from 10am - 12pm GMT+1, please let me know! Need a mentor who can start ASAP.
+
+                                    *NOTE: The budget is actually $40/hr
+
+                                    If the timeslot doesn't work, please indicate so.
+                                    [Optional: Mentors experienced with A, B, C secondary skills are highly preferred.]
+                                </li>
+                                <li>Check the I am looking for a long-term mentor box </li>
+                                <li>For the budget, choose $15/15 min, then click Send Request </li>
+                                <li>Relevant mentors will then be notified and show interest, leave a note, or message the user.
+                                    Interested mentors will appear at the bottom of the request for reference. </li>
+                                <li>Fellows should review and reach out to their preferred mentors; after reviewing their profiles.</li>
+                                <li>If a Fellow encounters any issues, the Success Operations Team  should be contacted via Mentorship Concerns channel on slack.
+                                </li>
+                            </ol>
+
+                            <b>Once a match is found</b>
+                            <ol type="1">
+                                <li>Notify the Success Operations Team via email
+                                    (mentorshiprequest@andela.com) about which mentor they have agreed to work with
+                                    and the details of their mentorship in the below format.
+                                    Please note that a request MUST follow this format.
+                                </li>
+                                <ol>
+                                    <p> <i>Proposed mentor: XXX </i></p>
+                                    <p> <i>Primary Skills: AngularJS, Typescript.</i></p>
+                                    <p> <i>Secondary Skills: Firebase </i></p>
+                                    <p> <i>Available time for pairing:  10:30 am-12:30 pm  GMT+3 (Monday- Friday)</i></p>
+                                    <p> <i>Length of mentorship: 3 Months</i></p>
+                                    <p> <i>A brief overview of the scope of work: The client uses Angular 1 as the main frontend
+                                            framework integrated with Typescript on the backend.
+                                        </i>
+                                    </p>
+                                    <p><i>Start Date: Mon 8th May</i></p>
+                                </ol>
+                                <li>The Success Operations Team  will review and vet the mentor and either approve or deny a request.
+                                    A reason for any deny will be provided.
+                                </li>
+                                <li>The Andela Codementor personnel would seek to close out on the request and have the mentor
+                                    sign the NDA and confirm all of the details that were agreed upon:
+                                </li>
+                                <ul>
+                                    <li>$40/hr rate
+                                    </li>
+                                    <li>daily/weekly commitment and time of commitment
+                                    </li>
+                                    <li>expected duration of mentorship
+                                    </li>
+                                </ul>
+                                <li>The Andela Codementor personnel will send out an intro email and schedule their sessions.
+                                </li>
+                            </ol>
+                            <!-- Body -->
+
+                            <!-- Footer section -->
+                            <table align="center">
+                                <tbody>
+                                <tr>
+                                    <th style="padding-top: 25px;">
+                                        <p style="font-size: 12px; color: #b3b3b3; font-weight: 100; text-align: center;">
+                                            This email was sent by the Andela Lenken Development Team.
+                                        </p>
+                                    </th>
+                                </tr>
+                                </tbody>
+                            </table>
+                            <!-- Footer section -->
+                        </td>
+                    </tr>
+                    </tbody>
+                </table>
+                <!-- Container -->
+            </center>
+        </td>
+    </tr>
+    </tbody>
+</table>
+<!-- Main table wrapper -->
+</body>
+</html>